backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python-headless
scikit-learn
scikit-image
rasterio
motor
pydantic
python-dotenv
matplotlib
pillow
